module3-task1: * merge app helpers
