--- src/helpers/number/get-random-number.helper.ts
@@ -1,7 +1,0 @@
-const getRandomNumber = (min: number, max: number) => {
-  const randomNumber = Math.floor(Math.random() * (max - min + 1)) + min;
-
-  return randomNumber;
-};
-
-export { getRandomNumber };
