+++ src/helpers/number/number.ts
@@ -1,1 +1,7 @@
-export * from './get-random-number.helper';
+const getRandomNumber = (min: number, max: number) => {
+  const randomNumber = Math.floor(Math.random() * (max - min + 1)) + min;
+
+  return randomNumber;
+};
+
+export { getRandomNumber };
